Consolidate target grader boilerplate into test-fixture.ts and update template & prompts
