--- guides/css/size-aware-styling/targets/daily-grind/patches/claude-solution.patch
@@ -1,168 +1,0 @@
-diff --git a/index.html b/index.html
-index 66c9173..f844927 100644
---- a/index.html
-+++ b/index.html
-@@ -107,12 +107,43 @@
-       margin-top: 20px;
-     }
- 
-+    /*
-+     * Card component: each card is wrapped in a container-query container so
-+     * its layout responds to the space it is given (e.g. in the grid above
-+     * vs. a narrow sidebar), not the viewport width.
-+     */
-+    .card-container {
-+      /* MANDATORY: marks this element as a query container for its children. */
-+      container-type: inline-size;
-+      container-name: card;
-+    }
-+
-+    /* Safe default layout: stacked, used whenever container queries (or their
-+       fallbacks below) don't apply. */
-     .card {
-       background: white;
-       padding: 20px;
-       border-radius: 8px;
-       box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-       text-align: center;
-+      display: flex;
-+      flex-direction: column;
-+      align-items: center;
-+      gap: 0.75rem;
-+      height: 100%;
-+    }
-+
-+    .card-icon {
-+      font-size: 2.5rem;
-+      line-height: 1;
-+      flex-shrink: 0;
-+    }
-+
-+    .card-body {
-+      display: flex;
-+      flex-direction: column;
-+      align-items: center;
-+      flex: 1;
-     }
- 
-     .card h3 {
-@@ -120,6 +151,68 @@
-       color: var(--primary);
-     }
- 
-+    /* Extra detail only shown when there is plenty of room. */
-+    .card-details {
-+      display: none;
-+      margin-top: 10px;
-+      padding-top: 10px;
-+      border-top: 1px solid var(--secondary);
-+      font-size: 0.85rem;
-+    }
-+
-+    /* Fallback for browsers without container query support: fall back to
-+       viewport width to decide when there is "probably" enough room. */
-+    @media (min-width: 700px) {
-+      .card {
-+        flex-direction: row;
-+        text-align: left;
-+        align-items: center;
-+      }
-+
-+      .card-body {
-+        align-items: flex-start;
-+      }
-+    }
-+
-+    /* Progressive enhancement: where container queries are supported, let the
-+       card's own width (not the viewport) drive the layout. */
-+    @supports (container-type: inline-size) {
-+      .card {
-+        /* Reset to the safe stacked default; @container rules below take over. */
-+        flex-direction: column;
-+        text-align: center;
-+        align-items: center;
-+      }
-+
-+      .card-body {
-+        align-items: center;
-+      }
-+
-+      /* Side-by-side layout once the card itself has room. */
-+      @container card (min-width: 400px) {
-+        .card {
-+          flex-direction: row;
-+          text-align: left;
-+          align-items: center;
-+        }
-+
-+        .card-body {
-+          align-items: flex-start;
-+        }
-+
-+        .card-icon {
-+          font-size: 3rem;
-+        }
-+      }
-+
-+      /* Show extra detail once the card is wide enough to fit it comfortably. */
-+      @container card (min-width: 600px) {
-+        .card-details {
-+          display: block;
-+        }
-+      }
-+    }
-+
-     .btn {
-       display: inline-block;
-       background: var(--primary);
-@@ -178,20 +271,38 @@
-     <section>
-       <h2 style="text-align: center;">Seasonal Favorites</h2>
-       <div class="grid">
--        <div class="card">
--          <h3>Maple Oat Latte</h3>
--          <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
--          <a href="/menu/seasonal" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <div class="card-icon" aria-hidden="true">🍁</div>
-+            <div class="card-body">
-+              <h3>Maple Oat Latte</h3>
-+              <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
-+              <a href="/menu/seasonal" class="btn">View Details</a>
-+              <p class="card-details">Pulled with our single-origin espresso and topped with a dusting of cinnamon.</p>
-+            </div>
-+          </div>
-         </div>
--        <div class="card">
--          <h3>Dark Roast Espresso</h3>
--          <p>Bold, smokey, and perfect for a morning boost.</p>
--          <a href="/menu/espresso" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <div class="card-icon" aria-hidden="true">☕</div>
-+            <div class="card-body">
-+              <h3>Dark Roast Espresso</h3>
-+              <p>Bold, smokey, and perfect for a morning boost.</p>
-+              <a href="/menu/espresso" class="btn">View Details</a>
-+              <p class="card-details">Roasted in small batches for a rich, low-acidity finish with notes of dark chocolate.</p>
-+            </div>
-+          </div>
-         </div>
--        <div class="card">
--          <h3>Cold Brew</h3>
--          <p>Steeped for 18 hours for maximum smoothness.</p>
--          <a href="/menu/cold-brew" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <div class="card-icon" aria-hidden="true">🧊</div>
-+            <div class="card-body">
-+              <h3>Cold Brew</h3>
-+              <p>Steeped for 18 hours for maximum smoothness.</p>
-+              <a href="/menu/cold-brew" class="btn">View Details</a>
-+              <p class="card-details">Slow-steeped and never bitter, served over ice with your choice of milk.</p>
-+            </div>
-+          </div>
-         </div>
-       </div>
-     </section>
