--- guides/css/size-aware-styling/targets/daily-grind/patches/codex-solution.patch
@@ -1,193 +1,0 @@
-diff --git a/index.html b/index.html
-index 66c9173..66bad00 100644
---- a/index.html
-+++ b/index.html
-@@ -102,17 +102,39 @@
- 
-     .grid {
-       display: grid;
--      grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-+      grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-       gap: 20px;
-       margin-top: 20px;
-     }
- 
-+    .card-container {
-+      container-type: inline-size;
-+    }
-+
-     .card {
-       background: white;
-       padding: 20px;
-       border-radius: 8px;
-       box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-       text-align: center;
-+      display: flex;
-+      flex-direction: column;
-+      gap: 16px;
-+      height: 100%;
-+    }
-+
-+    .card-media {
-+      min-height: 120px;
-+      border-radius: 6px;
-+      background-size: cover;
-+      background-position: center;
-+    }
-+
-+    .card-content {
-+      display: flex;
-+      flex: 1;
-+      flex-direction: column;
-+      align-items: center;
-     }
- 
-     .card h3 {
-@@ -120,6 +142,13 @@
-       color: var(--primary);
-     }
- 
-+    .card-details {
-+      display: none;
-+      margin-top: 10px;
-+      font-size: 0.92rem;
-+      color: #5d4037;
-+    }
-+
-     .btn {
-       display: inline-block;
-       background: var(--primary);
-@@ -135,6 +164,70 @@
-       background: #3e2723;
-     }
- 
-+    /* Fallback for browsers without container query support. */
-+    @media (min-width: 700px) {
-+      .card {
-+        flex-direction: row;
-+        text-align: left;
-+      }
-+
-+      .card-media {
-+        flex: 0 0 145px;
-+        min-height: 145px;
-+      }
-+
-+      .card-content {
-+        align-items: flex-start;
-+      }
-+
-+      .card-details {
-+        display: block;
-+      }
-+    }
-+
-+    @supports (container-type: inline-size) {
-+      @media (min-width: 700px) {
-+        .card {
-+          flex-direction: column;
-+          text-align: center;
-+        }
-+
-+        .card-media {
-+          flex: initial;
-+          min-height: 120px;
-+        }
-+
-+        .card-content {
-+          align-items: center;
-+        }
-+
-+        .card-details {
-+          display: none;
-+        }
-+      }
-+
-+      @container (min-width: 400px) {
-+        .card {
-+          flex-direction: row;
-+          align-items: stretch;
-+          text-align: left;
-+        }
-+
-+        .card-media {
-+          flex: 0 0 150px;
-+          min-height: 150px;
-+        }
-+
-+        .card-content {
-+          align-items: flex-start;
-+        }
-+
-+        .card-details {
-+          display: block;
-+        }
-+      }
-+    }
-+
-     footer {
-       background: var(--secondary);
-       text-align: center;
-@@ -178,20 +271,38 @@
-     <section>
-       <h2 style="text-align: center;">Seasonal Favorites</h2>
-       <div class="grid">
--        <div class="card">
--          <h3>Maple Oat Latte</h3>
--          <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
--          <a href="/menu/seasonal" class="btn">View Details</a>
-+        <div class="card-container">
-+          <article class="card">
-+            <div class="card-media" style="background-image: url('https://images.unsplash.com/photo-1461023058943-07fcbe16d735?auto=format&fit=crop&w=600&q=80');"></div>
-+            <div class="card-content">
-+              <h3>Maple Oat Latte</h3>
-+              <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
-+              <p class="card-details">Available hot or iced with house-pulled espresso and a cinnamon finish.</p>
-+              <a href="/menu/seasonal" class="btn">View Details</a>
-+            </div>
-+          </article>
-         </div>
--        <div class="card">
--          <h3>Dark Roast Espresso</h3>
--          <p>Bold, smokey, and perfect for a morning boost.</p>
--          <a href="/menu/espresso" class="btn">View Details</a>
-+        <div class="card-container">
-+          <article class="card">
-+            <div class="card-media" style="background-image: url('https://images.unsplash.com/photo-1510707577719-ae7c14805e3a?auto=format&fit=crop&w=600&q=80');"></div>
-+            <div class="card-content">
-+              <h3>Dark Roast Espresso</h3>
-+              <p>Bold, smokey, and perfect for a morning boost.</p>
-+              <p class="card-details">A rotating single-origin roast dialed in daily for balanced crema.</p>
-+              <a href="/menu/espresso" class="btn">View Details</a>
-+            </div>
-+          </article>
-         </div>
--        <div class="card">
--          <h3>Cold Brew</h3>
--          <p>Steeped for 18 hours for maximum smoothness.</p>
--          <a href="/menu/cold-brew" class="btn">View Details</a>
-+        <div class="card-container">
-+          <article class="card">
-+            <div class="card-media" style="background-image: url('https://images.unsplash.com/photo-1517701604599-bb29b565090c?auto=format&fit=crop&w=600&q=80');"></div>
-+            <div class="card-content">
-+              <h3>Cold Brew</h3>
-+              <p>Steeped for 18 hours for maximum smoothness.</p>
-+              <p class="card-details">Served over ice, with optional vanilla cold foam or oat milk.</p>
-+              <a href="/menu/cold-brew" class="btn">View Details</a>
-+            </div>
-+          </article>
-         </div>
-       </div>
-     </section>
-diff --git a/package.json b/package.json
-index 2bd4d77..712b282 100644
---- a/package.json
-+++ b/package.json
-@@ -2,6 +2,7 @@
-   "name": "daily-grind",
-   "version": "1.0.0",
-   "scripts": {
-+    "build": "node -e \"require('fs').accessSync('index.html')\"",
-     "start": "pnpm dlx statikk --port ${PORT:-11222}"
-   },
-   "pnpm": {
