--- guides/css/size-aware-styling/targets/daily-grind/patches/jetski-solution.patch
@@ -1,138 +1,0 @@
-diff --git a/index.html b/index.html
-index 66c9173..877c98f 100644
---- a/index.html
-+++ b/index.html
-@@ -107,16 +107,26 @@
-       margin-top: 20px;
-     }
- 
-+    .card-container {
-+      container-type: inline-size;
-+    }
-+
-     .card {
-       background: white;
-       padding: 20px;
-       border-radius: 8px;
-       box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-       text-align: center;
-+      display: flex;
-+      flex-direction: column;
-+      gap: 1rem;
-+      align-items: center;
-+      justify-content: center;
-+      height: 100%;
-     }
- 
-     .card h3 {
--      margin-bottom: 10px;
-+      margin-bottom: 0;
-       color: var(--primary);
-     }
- 
-@@ -127,14 +137,43 @@
-       padding: 10px 20px;
-       text-decoration: none;
-       border-radius: 5px;
--      margin-top: 15px;
-+      margin-top: 0;
-       transition: background 0.3s;
-+      white-space: nowrap;
-     }
- 
-     .btn:hover {
-       background: #3e2723;
-     }
- 
-+    /* Fallback using media queries for older browsers */
-+    @media (min-width: 600px) {
-+      .card {
-+        flex-direction: row;
-+        justify-content: space-between;
-+        text-align: left;
-+      }
-+    }
-+
-+    /* Overwrite with container queries where supported */
-+    @supports (container-type: inline-size) {
-+      @media (min-width: 600px) {
-+        .card {
-+          flex-direction: column;
-+          text-align: center;
-+        }
-+      }
-+
-+      @container (min-width: 400px) {
-+        .card {
-+          flex-direction: row;
-+          justify-content: space-between;
-+          text-align: left;
-+          align-items: center;
-+        }
-+      }
-+    }
-+
-     footer {
-       background: var(--secondary);
-       text-align: center;
-@@ -171,27 +210,33 @@
-   <header class="hero">
-     <h1>Wake Up Your Senses</h1>
-     <p>Ethically sourced, locally roasted, delivered to your cup.</p>
--    <a href="/order-online" class="btn">Order Now</a>
-+    <a href="/order-online" class="btn" style="margin-top: 15px;">Order Now</a>
-   </header>
- 
-   <main class="container">
-     <section>
-       <h2 style="text-align: center;">Seasonal Favorites</h2>
-       <div class="grid">
--        <div class="card">
--          <h3>Maple Oat Latte</h3>
--          <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
--          <a href="/menu/seasonal" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <h3>Maple Oat Latte</h3>
-+            <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
-+            <a href="/menu/seasonal" class="btn">View Details</a>
-+          </div>
-         </div>
--        <div class="card">
--          <h3>Dark Roast Espresso</h3>
--          <p>Bold, smokey, and perfect for a morning boost.</p>
--          <a href="/menu/espresso" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <h3>Dark Roast Espresso</h3>
-+            <p>Bold, smokey, and perfect for a morning boost.</p>
-+            <a href="/menu/espresso" class="btn">View Details</a>
-+          </div>
-         </div>
--        <div class="card">
--          <h3>Cold Brew</h3>
--          <p>Steeped for 18 hours for maximum smoothness.</p>
--          <a href="/menu/cold-brew" class="btn">View Details</a>
-+        <div class="card-container">
-+          <div class="card">
-+            <h3>Cold Brew</h3>
-+            <p>Steeped for 18 hours for maximum smoothness.</p>
-+            <a href="/menu/cold-brew" class="btn">View Details</a>
-+          </div>
-         </div>
-       </div>
-     </section>
-diff --git a/package.json b/package.json
-index 2bd4d77..94920fa 100644
---- a/package.json
-+++ b/package.json
-@@ -2,7 +2,8 @@
-   "name": "daily-grind",
-   "version": "1.0.0",
-   "scripts": {
--    "start": "pnpm dlx statikk --port ${PORT:-11222}"
-+    "start": "pnpm dlx statikk --port ${PORT:-11222}",
-+    "build": "echo Build succeeded"
-   },
-   "pnpm": {
-     "onlyBuiltDependencies": []
