--- guides/css/size-aware-styling/targets/devtools-times/patches/claude-solution.patch
@@ -1,107 +1,0 @@
-diff --git a/src/components/ArticleTeaser.astro b/src/components/ArticleTeaser.astro
-index af1fa02..b07b948 100644
---- a/src/components/ArticleTeaser.astro
-+++ b/src/components/ArticleTeaser.astro
-@@ -17,9 +17,9 @@ const isBookmarked = readingList.includes(article.id);
- const capitalizedCategory = article.data.category.charAt(0).toUpperCase() + article.data.category.slice(1);
- ---
- {large ? (
--    <div class="bg-white p-6 md:p-8 shadow-sm">
--        <div class="grid grid-cols-1 lg:grid-cols-3 gap-8">
--            <div class="lg:col-span-1 flex flex-col">
-+    <div class="feature-container bg-white p-6 md:p-8 shadow-sm">
-+        <div class="feature-card gap-8">
-+            <div class="feature-content flex flex-col">
-                 <a href={`${import.meta.env.BASE_URL}category/${article.data.category.toLowerCase()}`} class="text-amber-800 text-sm font-brand tracking-widest">{capitalizedCategory}</a>
-                 <a href={`${import.meta.env.BASE_URL}articles/${article.id}`} class="flex-grow">
-                     <h1 class="mt-2 text-3xl md:text-4xl font-brand text-gray-900 leading-tight hover:text-amber-400 transition-colors">
-@@ -37,10 +37,10 @@ const capitalizedCategory = article.data.category.charAt(0).toUpperCase() + arti
-                     <BookmarkButton client:visible slug={article.id} initialIsBookmarked={isBookmarked} />
-                 </div>
-             </div>
--            <div class="lg:col-span-2">
-+            <div class="feature-media">
-                 <a href={`${import.meta.env.BASE_URL}articles/${article.id}`}>
--                    <Image 
--                        src={image} 
-+                    <Image
-+                        src={image}
-                         alt=""
-                         width={800}
-                         height={533}
-@@ -80,4 +80,72 @@ const capitalizedCategory = article.data.category.charAt(0).toUpperCase() + arti
-             </div>
-         </div>
-     </article>
--)}
-\ No newline at end of file
-+)}
-+
-+<style>
-+    /* Mark the feature wrapper as a query container so `.feature-card`
-+       can react to the space it is actually given, not the viewport. */
-+    .feature-container {
-+        container-type: inline-size;
-+        container-name: feature;
-+    }
-+
-+    /* Safe default: stacked layout. Works even without container query
-+       support, so this must look correct on its own. */
-+    .feature-card {
-+        display: flex;
-+        flex-direction: column;
-+    }
-+
-+    .feature-media img {
-+        width: 100%;
-+        height: auto;
-+    }
-+
-+    /* Fallback for browsers without container queries: approximate the
-+       breakpoint using the viewport width instead. */
-+    @media (min-width: 1024px) {
-+        .feature-card {
-+            flex-direction: row;
-+        }
-+
-+        .feature-content {
-+            flex: 1 1 0%;
-+        }
-+
-+        .feature-media {
-+            flex: 2 1 0%;
-+        }
-+    }
-+
-+    /* Where supported, let the feature card's own width decide the
-+       layout so it still works when embedded in a narrower column. */
-+    @supports (container-type: inline-size) {
-+        .feature-card {
-+            flex-direction: column;
-+        }
-+
-+        .feature-content {
-+            flex: initial;
-+        }
-+
-+        .feature-media {
-+            flex: initial;
-+        }
-+
-+        @container feature (min-width: 700px) {
-+            .feature-card {
-+                flex-direction: row;
-+                align-items: stretch;
-+            }
-+
-+            .feature-content {
-+                flex: 1 1 0%;
-+            }
-+
-+            .feature-media {
-+                flex: 2 1 0%;
-+            }
-+        }
-+    }
-+</style>
-\ No newline at end of file
