--- guides/css/size-aware-styling/targets/devtools-times/patches/jetski-solution.patch
@@ -1,127 +1,0 @@
-diff --git a/src/components/ArticleTeaser.astro b/src/components/ArticleTeaser.astro
-index af1fa02..af7189f 100644
---- a/src/components/ArticleTeaser.astro
-+++ b/src/components/ArticleTeaser.astro
-@@ -52,32 +52,93 @@ const capitalizedCategory = article.data.category.charAt(0).toUpperCase() + arti
-         </div>
-     </div>
- ) : (
--    <article class="flex flex-col h-full group">
--        <a href={`${import.meta.env.BASE_URL}articles/${article.id}`}>
--            <Image 
--                src={image} 
--                alt={article.data.title} 
--                width={400}
--                height={225}
--                format="webp"
--                class="w-full object-cover rounded-md flex-shrink-0 aspect-[16/9]"
--            />
--        </a>
--        <div class="flex flex-col flex-grow mt-3">
--            <div class="text-amber-800 text-sm font-brand tracking-widest">{capitalizedCategory}</div>
--            <a href={`${import.meta.env.BASE_URL}articles/${article.id}`} class="flex flex-col">
--                <h3 class="text-lg md:text-xl font-brand text-gray-800 group-hover:text-amber-400 transition-colors flex-grow">
--                    {article.data.title}
--                </h3>
--                <p class="text-gray-600 mt-1">{article.data.summary}</p>
-+    <div class="card-container">
-+        <article class="card flex flex-col h-full group">
-+            <a href={`${import.meta.env.BASE_URL}articles/${article.id}`} class="card-image-link">
-+                <Image 
-+                    src={image} 
-+                    alt={article.data.title} 
-+                    width={400}
-+                    height={225}
-+                    format="webp"
-+                    class="card-image w-full object-cover rounded-md flex-shrink-0 aspect-[16/9]"
-+                />
-             </a>
--            <div class="flex justify-between text-gray-500 mt-2 pt-2 border-t-1 border-gray-100">
--                <span class="text-sm font-medium flex items-center">
--                    <BookOpenIcon className="w-4 h-4 inline-block mr-1 -mt-px" />
--                    {article.data.readingTime} Min.
--                </span>
--                <BookmarkButton client:visible slug={article.id} initialIsBookmarked={isBookmarked} />
-+            <div class="card-content flex flex-col flex-grow mt-3">
-+                <div class="text-amber-800 text-sm font-brand tracking-widest">{capitalizedCategory}</div>
-+                <a href={`${import.meta.env.BASE_URL}articles/${article.id}`} class="flex flex-col">
-+                    <h3 class="text-lg md:text-xl font-brand text-gray-800 group-hover:text-amber-400 transition-colors flex-grow">
-+                        {article.data.title}
-+                    </h3>
-+                    <p class="text-gray-600 mt-1">{article.data.summary}</p>
-+                </a>
-+                <div class="flex justify-between text-gray-500 mt-2 pt-2 border-t-1 border-gray-100">
-+                    <span class="text-sm font-medium flex items-center">
-+                        <BookOpenIcon className="w-4 h-4 inline-block mr-1 -mt-px" />
-+                        {article.data.readingTime} Min.
-+                    </span>
-+                    <BookmarkButton client:visible slug={article.id} initialIsBookmarked={isBookmarked} />
-+                </div>
-             </div>
--        </div>
--    </article>
--)}
-\ No newline at end of file
-+        </article>
-+    </div>
-+)}
-+
-+<style>
-+.card-container {
-+  container-type: inline-size;
-+}
-+
-+/* Default styles for small containers (stacked layout) */
-+.card {
-+  display: flex;
-+  flex-direction: column;
-+  gap: 1rem;
-+}
-+
-+/* Fallback using media queries for older browsers */
-+@media (min-width: 600px) {
-+  .card {
-+    flex-direction: row;
-+    align-items: center;
-+  }
-+  .card-image-link {
-+    width: 40%;
-+    flex-shrink: 0;
-+  }
-+  .card-content {
-+    margin-top: 0;
-+  }
-+}
-+
-+/* Overwrite with container queries where supported */
-+@supports (container-type: inline-size) {
-+  @media (min-width: 600px) {
-+    .card {
-+      /* Reset media query fallback if needed, or let container query handle it */
-+      flex-direction: column;
-+      align-items: stretch;
-+    }
-+    .card-image-link {
-+      width: 100%;
-+    }
-+    .card-content {
-+      margin-top: 0.75rem;
-+    }
-+  }
-+  
-+  @container (min-width: 400px) {
-+    .card {
-+      flex-direction: row;
-+      align-items: center;
-+    }
-+    .card-image-link {
-+      width: 40%;
-+      flex-shrink: 0;
-+    }
-+    .card-content {
-+      margin-top: 0;
-+    }
-+  }
-+}
-+</style>
-\ No newline at end of file
