+++ guides/test-fixture.ts
@@ -3,6 +3,12 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { spawn, spawnSync } from 'child_process';
 import * as net from 'net';
+import { fileURLToPath } from 'url';
+import { parseHTML } from 'linkedom';
+import { Project } from 'ts-morph';
+import { extractTargetFilesFromPatch } from '../lib/patch-utils.ts';
+
+export { expect } from '@playwright/test';
 
 export type ServerWorkerFixtures = {
   TARGET_URL: string;
@@ -115,4 +121,111 @@ export const test = base.extend<{}, ServerWorkerFixtures>({
   }, { scope: 'worker', timeout: 60000 }]
 });
 
-export { expect } from '@playwright/test';
+const BASE_APP_DEFAULT_FILES: Record<string, string[]> = {
+  'daily-grind': ['index.html'],
+  'devtools-times': [
+    'src/components/ArticleTeaser.astro',
+    'src/layouts/Layout.astro',
+    'src/styles/global.css',
+    'src/components/SearchFlyout.tsx',
+    'src/components/ReadingListFlyout.tsx',
+  ],
+};
+
+export function getTargetFiles(testFileUrl: string): string[] {
+  const patchFile = process.env.PATCH_FILE;
+  if (!patchFile) {
+    throw new Error('PATCH_FILE environment variable not set.');
+  }
+
+  const rootDir = process.cwd();
+  const baseAppName = path.basename(path.dirname(fileURLToPath(testFileUrl)));
+  const patchTargetFiles = extractTargetFilesFromPatch(patchFile);
+  const defaultFiles = BASE_APP_DEFAULT_FILES[baseAppName] || [];
+  const targetFiles = Array.from(new Set([...patchTargetFiles, ...defaultFiles]));
+  return targetFiles.map((f: string) => path.resolve(rootDir, f));
+}
+
+const HTML_EXTS = /\.(html|htm|astro)$/i;
+const CSS_EXTS = /\.css$/i;
+const JS_EXTS = /\.(js|ts|tsx|jsx)$/i;
+
+export function extractAllCss(files: string[]): string {
+  const cssBlocks: string[] = [];
+  for (const file of files) {
+    if (!fs.existsSync(file) || fs.statSync(file).isDirectory()) continue;
+    const content = fs.readFileSync(file, 'utf8');
+
+    if (CSS_EXTS.test(file)) {
+      cssBlocks.push(content);
+    } else if (HTML_EXTS.test(file)) {
+      try {
+        const { document } = parseHTML(content);
+        document.querySelectorAll('style').forEach((style: any) => {
+          if (style.textContent) cssBlocks.push(style.textContent);
+        });
+        document.querySelectorAll('[style]').forEach((el: any) => {
+          const inlineStyle = el.getAttribute('style');
+          if (inlineStyle) cssBlocks.push(inlineStyle);
+        });
+      } catch {
+        const styleMatches = content.match(/<style[^>]*>([\s\S]*?)<\/style>/gi);
+        if (styleMatches) cssBlocks.push(...styleMatches);
+      }
+    } else if (JS_EXTS.test(file)) {
+      const styleMatches = content.match(/<style[^>]*>([\s\S]*?)<\/style>/gi);
+      if (styleMatches) {
+        for (const match of styleMatches) {
+          cssBlocks.push(match.replace(/^<style[^>]*>/i, '').replace(/<\/style>$/i, ''));
+        }
+      }
+    }
+  }
+  return cssBlocks.join('\n').replace(/\s+/g, ' ');
+}
+
+export function populateJsProject(project: Project, files: string[]): void {
+  for (const file of files) {
+    if (!fs.existsSync(file) || fs.statSync(file).isDirectory()) continue;
+    const content = fs.readFileSync(file, 'utf8');
+
+    if (JS_EXTS.test(file) && !HTML_EXTS.test(file)) {
+      project.createSourceFile(file, content, { overwrite: true });
+    } else if (HTML_EXTS.test(file)) {
+      try {
+        if (file.endsWith('.astro')) {
+          const frontmatter = content.match(/^---[\r\n]+([\s\S]*?)[\r\n]+---/);
+          if (frontmatter && frontmatter[1]) {
+            project.createSourceFile(`${file}_frontmatter.ts`, frontmatter[1], { overwrite: true });
+          }
+        }
+        const { document } = parseHTML(content);
+        document.querySelectorAll('script').forEach((script: any, idx: number) => {
+          if (script.textContent) {
+            project.createSourceFile(`${file}_script_${idx}.ts`, script.textContent, { overwrite: true });
+          }
+        });
+      } catch {
+        // Fallback for non-standard HTML fragments
+      }
+    }
+  }
+}
+
+export function getJsProject(files: string[]): Project {
+  const project = new Project({ useInMemoryFileSystem: true });
+  populateJsProject(project, files);
+  return project;
+}
+
+export function getHtmlDocuments(files: string[]): Array<{ file: string; document: any }> {
+  const docs: Array<{ file: string; document: any }> = [];
+  for (const file of files) {
+    if (!fs.existsSync(file) || fs.statSync(file).isDirectory()) continue;
+    if (HTML_EXTS.test(file)) {
+      const content = fs.readFileSync(file, 'utf8');
+      docs.push({ file, document: parseHTML(content).document });
+    }
+  }
+  return docs;
+}
