--- guides/ui-behaviors/animate-to-from-top-layer/targets/daily-grind/patches/claude-solution.patch
@@ -1,223 +1,0 @@
-diff --git a/index.html b/index.html
-index 66c9173..f53c4be 100644
---- a/index.html
-+++ b/index.html
-@@ -152,6 +152,154 @@
-       text-decoration: none;
-       font-size: 0.85rem;
-     }
-+
-+    /* Hero action buttons */
-+    .hero-actions {
-+      display: flex;
-+      gap: 15px;
-+      margin-top: 15px;
-+      flex-wrap: wrap;
-+      justify-content: center;
-+    }
-+
-+    .btn-secondary {
-+      background: transparent;
-+      border: 1px solid white;
-+      color: white;
-+      font: inherit;
-+      cursor: pointer;
-+    }
-+
-+    .btn-secondary:hover {
-+      background: rgba(255, 255, 255, 0.15);
-+    }
-+
-+    /* --- Top-layer content (dialog + popover) --- */
-+
-+    #rewards-dialog {
-+      border: none;
-+      border-radius: 10px;
-+      padding: 30px;
-+      max-width: 400px;
-+      width: 90%;
-+      color: var(--text);
-+      box-shadow: 0 10px 40px rgba(0, 0, 0, 0.25);
-+    }
-+
-+    #rewards-dialog h2 {
-+      color: var(--primary);
-+      margin-bottom: 10px;
-+    }
-+
-+    #rewards-dialog p {
-+      margin-bottom: 15px;
-+    }
-+
-+    #rewards-dialog label {
-+      display: block;
-+      font-size: 0.9rem;
-+      margin-bottom: 20px;
-+    }
-+
-+    #rewards-dialog input {
-+      display: block;
-+      width: 100%;
-+      margin-top: 6px;
-+      padding: 8px 10px;
-+      border: 1px solid var(--secondary);
-+      border-radius: 4px;
-+      font: inherit;
-+    }
-+
-+    .dialog-actions {
-+      display: flex;
-+      justify-content: flex-end;
-+      gap: 10px;
-+    }
-+
-+    .dialog-actions .btn {
-+      margin-top: 0;
-+    }
-+
-+    .dialog-actions .btn-secondary {
-+      background: transparent;
-+      border: 1px solid var(--primary);
-+      color: var(--primary);
-+    }
-+
-+    #hours-popover {
-+      border: none;
-+      border-radius: 8px;
-+      padding: 20px;
-+      max-width: 260px;
-+      color: var(--text);
-+      box-shadow: 0 6px 24px rgba(0, 0, 0, 0.2);
-+    }
-+
-+    #hours-popover h3 {
-+      color: var(--primary);
-+      margin-bottom: 8px;
-+    }
-+
-+    /* 1 & 3. Visible (open) state + entry starting styles */
-+    dialog[open],
-+    [popover]:popover-open {
-+      opacity: 1;
-+      transform: scale(1);
-+
-+      @starting-style {
-+        opacity: 0;
-+        transform: scale(0.9);
-+      }
-+    }
-+
-+    /* Base (closed/exit) state and transitions */
-+    dialog,
-+    [popover] {
-+      opacity: 0;
-+      transform: scale(0.9);
-+
-+      /* MANDATORY: transition display and overlay for top-layer elements */
-+      transition-property: opacity, transform, display, overlay;
-+      transition-duration: 0.3s;
-+      transition-timing-function: ease-out;
-+      /* Applies to discrete properties like display and overlay */
-+      transition-behavior: allow-discrete;
-+    }
-+
-+    /* 4. Animate the dialog's backdrop */
-+    dialog::backdrop {
-+      background-color: rgba(0, 0, 0, 0);
-+      transition:
-+        display 0.3s allow-discrete,
-+        overlay 0.3s allow-discrete,
-+        background-color 0.3s ease-out;
-+    }
-+
-+    dialog[open]::backdrop {
-+      background-color: rgba(0, 0, 0, 0.5);
-+
-+      @starting-style {
-+        background-color: rgba(0, 0, 0, 0);
-+      }
-+    }
-+
-+    /* 5. Respect user preference for reduced motion */
-+    @media (prefers-reduced-motion: reduce) {
-+
-+      dialog,
-+      [popover] {
-+        transform: none;
-+        transition-duration: 0.1s;
-+      }
-+
-+      dialog[open],
-+      [popover]:popover-open {
-+        @starting-style {
-+          transform: none;
-+        }
-+      }
-+    }
-   </style>
- </head>
- 
-@@ -171,9 +319,34 @@
-   <header class="hero">
-     <h1>Wake Up Your Senses</h1>
-     <p>Ethically sourced, locally roasted, delivered to your cup.</p>
--    <a href="/order-online" class="btn">Order Now</a>
-+    <div class="hero-actions">
-+      <a href="/order-online" class="btn">Order Now</a>
-+      <button id="open-rewards-dialog" class="btn btn-secondary" type="button">Join Rewards</button>
-+      <button popovertarget="hours-popover" class="btn btn-secondary" type="button">Store Hours</button>
-+    </div>
-   </header>
- 
-+  <dialog id="rewards-dialog">
-+    <form method="dialog">
-+      <h2>Join The Daily Grind Rewards</h2>
-+      <p>Sign up today and get a free drink on your birthday, plus a discount on your first order.</p>
-+      <label>
-+        Email address
-+        <input type="email" name="email" required placeholder="you@example.com">
-+      </label>
-+      <div class="dialog-actions">
-+        <button value="cancel" class="btn btn-secondary" type="submit">Cancel</button>
-+        <button value="confirm" class="btn" type="submit">Sign Up</button>
-+      </div>
-+    </form>
-+  </dialog>
-+
-+  <div id="hours-popover" popover>
-+    <h3>Today's Hours</h3>
-+    <p>Mon&ndash;Fri: 6am &ndash; 7pm</p>
-+    <p>Sat&ndash;Sun: 7am &ndash; 6pm</p>
-+  </div>
-+
-   <main class="container">
-     <section>
-       <h2 style="text-align: center;">Seasonal Favorites</h2>
-@@ -207,6 +380,27 @@
-     <p>&copy; 2026 The Daily Grind Coffee Co. All rights reserved.</p>
-   </footer>
- 
-+  <script>
-+    // Feature detection for native top-layer animation support.
-+    const supportsTopLayerAnimation =
-+      window.CSS &&
-+      CSS.supports('transition-behavior', 'allow-discrete') &&
-+      CSS.supports('overlay', 'auto');
-+
-+    const rewardsDialog = document.getElementById('rewards-dialog');
-+    document.getElementById('open-rewards-dialog').addEventListener('click', () => {
-+      rewardsDialog.showModal();
-+    });
-+
-+    if (!supportsTopLayerAnimation) {
-+      // Manual JS fallback: skip the transition and let the dialog/popover
-+      // toggle instantly, since @starting-style/allow-discrete/overlay
-+      // are unavailable in this browser.
-+      rewardsDialog.style.transitionDuration = '0s';
-+      document.getElementById('hours-popover').style.transitionDuration = '0s';
-+    }
-+  </script>
-+
- </body>
- 
- </html>
