--- guides/ui-behaviors/animate-to-from-top-layer/targets/daily-grind/patches/jetski-solution.patch
@@ -1,216 +1,0 @@
-diff --git a/index.html b/index.html
-index 66c9173..70def72 100644
---- a/index.html
-+++ b/index.html
-@@ -152,6 +152,136 @@
-       text-decoration: none;
-       font-size: 0.85rem;
-     }
-+
-+    /* Dialog and Popover Layout */
-+    dialog {
-+      margin: auto;
-+      padding: 2rem;
-+      border: 1px solid var(--secondary);
-+      border-radius: 8px;
-+      background: white;
-+      color: var(--text);
-+      max-width: 500px;
-+      width: 90%;
-+      box-shadow: 0 10px 25px rgba(0, 0, 0, 0.2);
-+    }
-+
-+    [popover] {
-+      margin: auto;
-+      padding: 1.5rem;
-+      border: 1px solid var(--secondary);
-+      border-radius: 8px;
-+      background: white;
-+      color: var(--text);
-+      max-width: 350px;
-+      box-shadow: 0 5px 15px rgba(0, 0, 0, 0.15);
-+    }
-+
-+    .popover-btn {
-+      background: transparent;
-+      border: 1px solid var(--accent);
-+      color: white;
-+      padding: 5px 12px;
-+      border-radius: 4px;
-+      cursor: pointer;
-+      font-size: 0.9rem;
-+      transition: background 0.3s;
-+    }
-+
-+    .popover-btn:hover {
-+      background: var(--accent);
-+      color: var(--primary);
-+    }
-+
-+    /* Base (closed/exit) state & transitions */
-+    dialog,
-+    [popover] {
-+      opacity: 0;
-+      transform: scale(0.9);
-+
-+      transition:
-+        opacity 0.3s ease-out,
-+        transform 0.3s ease-out,
-+        display 0.3s allow-discrete,
-+        overlay 0.3s allow-discrete;
-+      transition-property: opacity, transform, display, overlay;
-+      transition-duration: 0.3s;
-+      transition-timing-function: ease-out;
-+      transition-behavior: allow-discrete;
-+    }
-+
-+    /* Visible (open) state */
-+    dialog[open],
-+    [popover]:popover-open {
-+      opacity: 1;
-+      transform: scale(1);
-+
-+      /* Entry starting state */
-+      @starting-style {
-+        opacity: 0;
-+        transform: scale(0.9);
-+      }
-+    }
-+
-+    @starting-style {
-+      dialog[open],
-+      [popover]:popover-open {
-+        opacity: 0;
-+        transform: scale(0.9);
-+      }
-+    }
-+
-+    /* Backdrop animations */
-+    dialog::backdrop,
-+    [popover]::backdrop {
-+      background-color: rgba(0, 0, 0, 0);
-+      transition:
-+        display 0.3s allow-discrete,
-+        overlay 0.3s allow-discrete,
-+        background-color 0.3s ease-out;
-+      transition-behavior: allow-discrete;
-+    }
-+
-+    dialog[open]::backdrop,
-+    [popover]:popover-open::backdrop {
-+      background-color: rgba(0, 0, 0, 0.5);
-+
-+      @starting-style {
-+        background-color: rgba(0, 0, 0, 0);
-+      }
-+    }
-+
-+    @starting-style {
-+      dialog[open]::backdrop,
-+      [popover]:popover-open::backdrop {
-+        background-color: rgba(0, 0, 0, 0);
-+      }
-+    }
-+
-+    /* Reduced Motion Support */
-+    @media (prefers-reduced-motion: reduce) {
-+      dialog,
-+      [popover] {
-+        transform: none;
-+        transition-duration: 0.1s;
-+      }
-+
-+      dialog[open],
-+      [popover]:popover-open {
-+        transform: none;
-+
-+        @starting-style {
-+          transform: none;
-+        }
-+      }
-+
-+      @starting-style {
-+        dialog[open],
-+        [popover]:popover-open {
-+          transform: none;
-+        }
-+      }
-+    }
-   </style>
- </head>
- 
-@@ -163,6 +293,7 @@
-       <li><a href="/menu">Menu</a></li>
-       <li><a href="/locations">Locations</a></li>
-       <li><a href="/rewards">Rewards</a></li>
-+      <li><button popovertarget="info-popover" class="popover-btn">Daily Offer</button></li>
-       <li><a href="/account">My Account</a></li>
-       <li><a href="/logout" class="logout-link">Logout</a></li>
-     </ul>
-@@ -171,7 +302,7 @@
-   <header class="hero">
-     <h1>Wake Up Your Senses</h1>
-     <p>Ethically sourced, locally roasted, delivered to your cup.</p>
--    <a href="/order-online" class="btn">Order Now</a>
-+    <button class="btn" onclick="document.getElementById('order-dialog').showModal()">Order Now</button>
-   </header>
- 
-   <main class="container">
-@@ -181,22 +312,40 @@
-         <div class="card">
-           <h3>Maple Oat Latte</h3>
-           <p>Creamy oat milk with a hint of Vermont maple syrup.</p>
--          <a href="/menu/seasonal" class="btn">View Details</a>
-+          <button onclick="document.getElementById('order-dialog').showModal()" class="btn">View Details</button>
-         </div>
-         <div class="card">
-           <h3>Dark Roast Espresso</h3>
-           <p>Bold, smokey, and perfect for a morning boost.</p>
--          <a href="/menu/espresso" class="btn">View Details</a>
-+          <button onclick="document.getElementById('order-dialog').showModal()" class="btn">View Details</button>
-         </div>
-         <div class="card">
-           <h3>Cold Brew</h3>
-           <p>Steeped for 18 hours for maximum smoothness.</p>
--          <a href="/menu/cold-brew" class="btn">View Details</a>
-+          <button onclick="document.getElementById('order-dialog').showModal()" class="btn">View Details</button>
-         </div>
-       </div>
-     </section>
-   </main>
- 
-+  <!-- Dialog Element -->
-+  <dialog id="order-dialog">
-+    <h2>Order Your Coffee</h2>
-+    <p>Select your favorite brew and customize size, milk, and sweetener choices.</p>
-+    <div style="margin-top: 20px; text-align: right;">
-+      <button class="btn" onclick="document.getElementById('order-dialog').close()">Close</button>
-+    </div>
-+  </dialog>
-+
-+  <!-- Popover Element -->
-+  <div id="info-popover" popover>
-+    <h3>Today's Special</h3>
-+    <p>Get 20% off all artisanal cold brews today only!</p>
-+    <div style="margin-top: 15px; text-align: right;">
-+      <button class="btn" popovertarget="info-popover" popovertargetaction="hide">Got it</button>
-+    </div>
-+  </div>
-+
-   <footer>
-     <div class="footer-links">
-       <a href="/about">Our Story</a> |
-diff --git a/package.json b/package.json
-index 2bd4d77..bb70d39 100644
---- a/package.json
-+++ b/package.json
-@@ -2,7 +2,8 @@
-   "name": "daily-grind",
-   "version": "1.0.0",
-   "scripts": {
--    "start": "pnpm dlx statikk --port ${PORT:-11222}"
-+    "start": "pnpm dlx statikk --port ${PORT:-11222}",
-+    "build": "echo 'Build successful'"
-   },
-   "pnpm": {
-     "onlyBuiltDependencies": []
