--- guides/ui-behaviors/animate-to-from-top-layer/targets/daily-grind/task.md
@@ -1,1 +1,0 @@
-- Add smooth entry and exit animations for top layer elements such as dialogs, popovers, and tooltips.
